+++ packages/react-static/src/static/getConfig.js
@@ -102,7 +102,6 @@ export const buildConfig = async (config = {}) => {
     renderToHtml: (render, comp) => render(comp),
     prefetchRate: 3,
     maxThreads: Infinity,
-    disableRouteInfoWarning: false,
     disableRoutePrefixing: false,
     outputFileRate: 100,
     extensions: ['.js', '.jsx'],
@@ -124,8 +123,6 @@ export const buildConfig = async (config = {}) => {
   // Set env variables to be used client side
   process.env.REACT_STATIC_MIN_LOAD_TIME = config.minLoadTime
   process.env.REACT_STATIC_PREFETCH_RATE = config.prefetchRate
-  process.env.REACT_STATIC_DISABLE_ROUTE_INFO_WARNING =
-    config.disableRouteInfoWarning
   process.env.REACT_STATIC_DISABLE_ROUTE_PREFIXING =
     config.disableRoutePrefixing